Related Entities

+++ src/Controller/ChimpokomonController.php
@@ -3,38 +3,32 @@
 namespace App\Controller;
 
 use App\Entity\Chimpokomon;
-use App\Repository\ChimpokomonRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use App\Repository\ChimpokomonRepository;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class ChimpokomonController extends AbstractController
 {
 
 
-    // #[Route('/chimpokomon', name: 'app_chimpokomon')]
-    // public function index(): JsonResponse
-    // {
-    //     return $this->json([
-    //         'message' => 'Welcome to your new controller!',
-    //         'path' => 'src/Controller/ChimpokomonController.php',
-    //     ]);
-    // }
-
-
-    #[Route('/chimpokomon', name: 'app_chimpokomon')]
+    #[Route('/chimpokomon', name: 'app_chimpokomon_getAll')]
     public function getAllChimpokomons(
-        ChimpokomonRepository $chimpokomonRepository
+        ChimpokomonRepository $chimpokomonRepository,
+        SerializerInterface $serializer
     ): JsonResponse
     {
-        $jsonChimpokos = $chimpokomonRepository->findAll();
-        return $this->json($jsonChimpokos);
+        $jsonChimpokos = $serializer->serialize($chimpokomonRepository->findStatusOn(), 'json',["groups" => "chimpokomon"]);
+        return new JsonResponse($jsonChimpokos, Response::HTTP_OK, [], true);
+        // return $this->json($jsonChimpokos);
     }
 
 
-    #[Route('/chimpokomon/{chimpokomon}', name: 'app_chimpokomon', methods:['GET'])]
+    #[Route('/chimpokomon/{chimpokomon}', name: 'app_chimpokomon_get', methods:['GET'])]
     public function getChimpokomon(
         Chimpokomon $chimpokomon
     ): JsonResponse
@@ -44,7 +38,7 @@ class ChimpokomonController extends AbstractController
 
 
 
-    #[Route('/chimpokomon/new', name:'app_chimpokon_new', methods:["POST"])]
+    #[Route('/chimpokomon', name:'app_chimpokon_create', methods:["POST"])]
       public function createChimpokomon(
         Request $request,
         EntityManagerInterface $entityManager
@@ -58,19 +52,44 @@ class ChimpokomonController extends AbstractController
         $entityManager->flush();
         return $this->json($newChimpo);
     }
-    #[Route('/chimpokomon/{chimpokomon}', name: 'app_chimpokomon', methods:['DELETE'])]
-    public function deleteChimpokomon(
+
+    #[Route('/chimpokomon/{chimpokomon}', name: 'app_chimpokomon_update', methods:['PUT', 'PATCH'])]
+      public function updateChimpokomon(
         Chimpokomon $chimpokomon,
+        Request $request,
         EntityManagerInterface $entityManager
     ): JsonResponse
     {
-        $entityManager->remove($chimpokomon);
-        $entityManager->flush();
 
-        return $this->json(null);
+
+        $chimpokomon->setName($request->toArray()['name'] ?? $chimpokomon->getName());
+        $entityManager->persist($chimpokomon);
+        $entityManager->flush();
+        return $this->json($chimpokomon);
     }
 
 
 
+
+    #[Route('/chimpokomon/{chimpokomon}', name: 'app_chimpokomon_delete', methods:['DELETE'])]
+    public function deleteChimpokomon(
+        Chimpokomon $chimpokomon,
+        EntityManagerInterface $entityManager,
+        Request $request
+    ): JsonResponse
+    {
+        $force = $request->toArray()["force"] ?? false;
+        if($force){
+            $entityManager->remove($chimpokomon);
+
+        }else{
+            $chimpokomon->setStatus('off');
+            $entityManager->persist($chimpokomon);
+        }
+
+        $entityManager->flush();
+
+        return $this->json(null);
+    }
 
 }
